Bug fix for userService delete method

+++ src/main/java/com/example/Conversa_API/services/UserService.java
@@ -1,7 +1,9 @@
 package com.example.Conversa_API.services;
 
+import com.example.Conversa_API.models.Message;
 import com.example.Conversa_API.models.User;
 import com.example.Conversa_API.models.UserDTO;
+import com.example.Conversa_API.repositories.MessageRepository;
 import com.example.Conversa_API.repositories.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -14,6 +16,8 @@ public class UserService {
 
     @Autowired
     UserRepository userRepository;
+    @Autowired
+    MessageRepository messageRepository;
 
 
     public List<User> getAllUsers() {
@@ -30,8 +34,15 @@ public class UserService {
     }
 
     public Optional<User> deleteUser(Long id) {
-        Optional<User> user =  userRepository.findById(id);
-        userRepository.deleteById(id);
-        return user;
+        Optional<User> userOptional = userRepository.findById(id);
+        if (userOptional.isPresent()){
+            User user = userOptional.get();
+            List<Message> messagesToDelete = user.getMessages();
+            for (Message message : messagesToDelete){
+                messageRepository.deleteById(message.getId());
+            }
+            userRepository.deleteById(id);
+        }
+        return userOptional;
     }
 }
